==> analysis/2-explore.Rmd <==
---
title: "1_import"
output: html_notebook
editor_options: 
  chunk_output_type: console
---
# Import library

```{r}
rm (list = ls())

# Helper
library (tidyverse)

# Import
library (rio)

# Exploration 
library (DataExplorer)
library (janitor)
library (arsenal)
library (viridis)

```

# Import

```{r}
df <- readRDS ("output/combined_data.RDS") 
  #filter (study != "lencioni") 

df <- df %>%
  mutate (age2 = case_when(
    age %in% c(3:5) ~ "<6yo",
    age %in% c(6:9) ~ "6-9yo",
    age %in% c(10:19) ~ "10+yo",
    age %in% c(20:29) ~ "20+yo",
    age %in% c(30:39) ~ "30+yo",
    age %in% c(40:49) ~ "40+yo",
    age %in% c(50:59) ~ "50+yo",
    age %in% c(60:69) ~ "60+yo",
    age %in% c(70:79) ~ "70+yo",
    TRUE ~ "80+yo"
  ))
```


# Plot

```{r}
unique(df$age)

df %>% 
  dplyr::select (subj, joint, study) %>%
  distinct() %>%
  group_by(joint, study) %>%
  summarise(n())
```


## Joint power by age category

```{r}

age_levels <- c("<6yo", "6-9yo","10+yo","20+yo","30+yo", "40+yo",
                "50+yo","60+yo","70+yo","80+yo")

age_labels <- c("<6", "6-9","10-19","20-29","30-39", "40-49",
                "50-59","60-69","70-79","80-89")
df_plot <- df %>%
  mutate (subj = factor (subj)) %>%
  mutate (joint = factor (joint, levels = c("ankle", "knee", "hip"))) %>%
  mutate (age2 = factor (age2, levels = age_levels, labels= age_labels)) %>%
  group_by(age2, joint, cycle) %>%
  summarise (power = mean (pow))

f0 <- df_plot %>%
  ggplot() +
  geom_line(aes(x = cycle, y = power, color = age2), linewidth = 0.8) +
  geom_hline(yintercept = 0, linetype = 2) + 
  scale_color_viridis(discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE, option = "D")+
  facet_wrap(~joint, scales = "free") +
  ylab ("Joint power (W/kg)") +
  xlab ("Stride (%)") +
  labs (color = "Age (yr)") +
  cowplot::theme_cowplot()


tiff ("manuscript/fig2.tiff", height = 3, width = 10, res = 100, unit = "in")
f0
dev.off()
```

```{r}
df_plot <- df_plot %>%
  filter (age2 == "30-39")

f0 <- df_plot %>%
  ggplot() +
  geom_line(aes(x = cycle, y = power), color = "black", linewidth = 0.8) +
  geom_hline(yintercept = 0, linetype = 2) + 
  facet_wrap(~joint, scales = "free", ncol = 3) +
  ylab ("Joint power (W/kg)") +
  xlab ("Stride (%)") +
  labs (color = "Age (yr)") +
  cowplot::theme_cowplot()

```


## Joint power by age category split by study

```{r}

df_plot <- df %>%
  mutate (subj = factor (subj)) %>%
  mutate (joint = factor (joint, levels = c("ankle", "knee", "hip"))) %>%
  mutate (age2 = factor (age2, levels = age_levels,  labels= age_labels)) %>%
  group_by(age2, joint, cycle, study) %>%
  summarise (power = mean (pow))

f0 <- df_plot %>%
  ggplot() +
  geom_line(aes(x = cycle, y = power, color = age2), linewidth = 0.8) +
  geom_hline(yintercept = 0, linetype = 2) + 
  scale_color_viridis(discrete = TRUE, option = "D")+
  scale_fill_viridis(discrete = TRUE, option = "D")+
  facet_wrap(study~joint, scales = "free", ncol = 3) +
  ylab ("Joint power (W/kg)") +
  xlab ("Stride (%)") +
  labs (color = "Age (yr)") +
  cowplot::theme_cowplot()


tiff ("manuscript/figSM1.tiff", height = 20, width = 15, res = 100, unit = "in")
f0
dev.off()
```

## Average power by age category split by study

```{r}

df_plot <- df %>%
  mutate (subj = factor (subj)) %>%
  mutate (joint = factor (joint, levels = c("ankle", "knee", "hip"))) %>%
  mutate (age2 = factor (age2, levels = age_levels,  labels= age_labels)) %>%
  group_by(age2, joint, study) %>%
  summarise (Mean = mean ( abs(pow)),
             Sd = sd(abs (pow)))

f0 <- df_plot %>%
  ggplot() +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.6)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.6), width = 0) +
  facet_wrap(~joint, scales = "free", ncol = 3) +
  xlab("Age category") +
  ylab ("Average power (W/kg)") +
  cowplot::theme_cowplot()



tiff ("manuscript/figSM1.tiff", height = 20, width = 15, res = 100, unit = "in")
f0
dev.off()
```

## Descriptors

```{r}
df_plot <- df %>%
  dplyr::select (subj, age, age2, sex, wt, ht, speed, strlen) %>%
  mutate (age2 = factor (age2, levels = age_levels, labels= age_labels)) 

age <- df_plot %>%
  dplyr::select (subj, age2, age) %>%
  distinct() %>%
  group_by(age2) %>%
  summarise (Mean = mean (age),
             Sd = sd (age)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) +
  xlab("Age category") +
  ylab ("Age (years)") +
  cowplot::theme_cowplot()

ht <- df_plot %>%
  dplyr::select (subj, age2, ht) %>%
  distinct() %>%
  group_by(age2) %>%
  summarise (Mean = mean (ht),
             Sd = sd (ht)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) +
  xlab("Age category") +
  ylab ("Height (m)") +
  cowplot::theme_cowplot()

wt <- df_plot %>%
  dplyr::select (subj, age2, wt) %>%
  distinct() %>%
  group_by(age2) %>%
  summarise (Mean = mean (wt),
             Sd = sd (wt)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) +
  xlab("Age category") +
  ylab ("Body mass (kg)") +
  cowplot::theme_cowplot()

speed <- df_plot %>%
  dplyr::select (subj, age2, speed) %>%
  distinct() %>%
  group_by(age2) %>%
  summarise (Mean = mean (speed),
             Sd = sd (speed)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) +
  xlab("Age category") +
  ylab ("Speed (m/s)") +
  cowplot::theme_cowplot()

strlen <- df_plot %>%
  dplyr::select (subj, age2, strlen) %>%
  distinct() %>%
  group_by(age2) %>%
  summarise (Mean = mean (strlen),
             Sd = sd (strlen)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd), width = 0) +
  xlab("Age category") +
  ylab ("Stride length (m)") +
  cowplot::theme_cowplot()

sex <- df_plot %>%
  dplyr::select (subj, age2, sex) %>%
  distinct() %>%
  group_by(age2, sex) %>%
  summarise (Mean = n()) %>%
  ggplot () +
  geom_bar (aes (x = age2, y = Mean, fill = sex), stat = "identity",
            position = "dodge") +
  xlab("Age category") +
  ylab ("Number of participants") +
  cowplot::theme_cowplot()

tiff ("manuscript/fig1.tiff", height = 10, width = 15, res = 100, unit = "in")
cowplot::plot_grid (age, ht, wt, 
           speed, strlen, sex, ncol = 2)
dev.off()
```


```{r}
t1 <- age$data%>%
  mutate (Mean = round (Mean, 2),
          Sd = round (Sd, 2)) %>%
  mutate ("Age (years)" = paste0(Mean, " (", Sd, ")")) %>%
  dplyr::select(-c(Mean, Sd))

t2 <- sex$data %>%
  dplyr::select(age2, sex, Mean) %>%
  pivot_wider(names_from = "sex",
              values_from = "Mean") %>%
  mutate ("Sex (female/male)" = paste0(f, "/", m)) %>%
  dplyr::select(-c(f, m))

t3 <- ht$data%>%
  mutate (Mean = round (Mean, 2),
          Sd = round (Sd, 2)) %>%
  mutate ("Height (m)" = paste0(Mean, " (", Sd, ")")) %>%
  dplyr::select(-c(Mean, Sd))

t4 <- wt$data%>%
  mutate (Mean = round (Mean, 2),
          Sd = round (Sd, 2)) %>%
  mutate ("Mass (kg)" = paste0(Mean, " (", Sd, ")")) %>%
  dplyr::select(-c(Mean, Sd))

t5 <- strlen$data %>%
  mutate (Mean = round (Mean, 2),
          Sd = round (Sd, 2)) %>%
  mutate ("Stride length (m)" = paste0(Mean, " (", Sd, ")")) %>%
  dplyr::select(-c(Mean, Sd))

t6 <- speed$data %>%
  mutate (Mean = round (Mean, 2),
          Sd = round (Sd, 2)) %>%
  mutate ("Speed (m/s)" = paste0(Mean, " (", Sd, ")")) %>%
  dplyr::select(-c(Mean, Sd))


descrp_df <- t1 %>%
  inner_join(t2, by = "age2") %>%
  inner_join(t3, by = "age2")%>%
  inner_join(t4, by = "age2")%>%
  inner_join(t5, by = "age2")%>%
  inner_join(t6, by = "age2") %>%
  rename ("Age category" = age2)

writexl::write_xlsx(descrp_df, "manuscript/sm_table.xlsx")

```

# Demography stratified by age

```{r}
df_plot <- df %>%
  dplyr::select (subj, study, age, age2, sex, wt, ht, speed, strlen) %>%
  mutate (age2 = factor (age2, levels = age_levels, labels= age_labels)) 

age_strat <- df %>%
  dplyr::select (subj, study, age, age2) %>%
  distinct() %>%
  group_by(study, age2)  %>%
  summarise (Mean = n()) %>%
  ggplot () +
  geom_bar (aes (x = age2, y = Mean, fill = study), stat = "identity",
            position = "dodge") +
  xlab("Age category") +
  ylab ("Number of participants") +
  cowplot::theme_cowplot()

age <- df_plot %>%
  dplyr::select (subj, study, age2, age) %>%
  distinct() %>%
  group_by(study, age2) %>%
  summarise (Mean = mean (age),
             Sd = sd (age)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.4)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.4), width = 0) +
  xlab("Age category") +
  ylab ("Age (years)") +
  cowplot::theme_cowplot()

ht <- df_plot %>%
  dplyr::select (subj, study, age2, ht) %>%
  distinct() %>%
  group_by(study, age2) %>%
  summarise (Mean = mean (ht),
             Sd = sd (ht)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.4)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.4), width = 0) +
  xlab("Age category") +
  ylab ("Height (m)") +
  cowplot::theme_cowplot()

wt <- df_plot %>%
  dplyr::select (subj, study, age2, wt) %>%
  distinct() %>%
  group_by(study, age2) %>%
  summarise (Mean = mean (wt),
             Sd = sd (wt)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.4)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.4), width = 0) +
  xlab("Age category") +
  ylab ("Body mass (kg)") +
  cowplot::theme_cowplot()

speed <- df_plot %>%
  dplyr::select (subj, study, age2, speed) %>%
  distinct() %>%
  group_by(study, age2) %>%
  summarise (Mean = mean (speed),
             Sd = sd (speed)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.4)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.4), width = 0) +
  xlab("Age category") +
  ylab ("Speed (m/s)") +
  cowplot::theme_cowplot()

strlen <- df_plot %>%
  dplyr::select (subj, study, age2, strlen) %>%
  distinct() %>%
  group_by(study, age2) %>%
  summarise (Mean = mean (strlen),
             Sd = sd (strlen)) %>%
  ggplot () +
  geom_point (aes (x = age2, y = Mean, color = study), 
              position = position_dodge(width = 0.4)) +
  geom_errorbar(aes (x = age2, ymin = Mean - Sd, ymax = Mean + Sd, color = study), 
                position = position_dodge(width = 0.4), width = 0) +
  xlab("Age category") +
  ylab ("Stride length (m)") +
  cowplot::theme_cowplot()

sex <- df_plot %>%
  dplyr::select (subj, study, age2, sex) %>%
  distinct() %>%
  mutate (sex = as.numeric(sex) - 1) %>%
  group_by(study, age2) %>%
  summarise (Mean = sum (sex)/n()) %>%
  ggplot () +
  geom_bar (aes (x = age2, y = Mean, fill = study), stat = "identity",
            position = "dodge") +
  xlab("Age category") +
  ylab ("Propotion of male participants") +
  cowplot::theme_cowplot()

tiff ("manuscript/fig SM2.tiff", height = 15, width = 15, res = 100, unit = "in")
cowplot::plot_grid (age_strat, age, ht, wt, 
           speed, strlen, sex, ncol = 2, labels = "auto")
dev.off()
```
