==> analysis/4-report_v2.Rmd <==
---
title: "3-model"
output: html_notebook
editor_options: 
  chunk_output_type: console
---
# Import library

```{r}
rm (list = ls())

# Helper
library (tidyverse)

# Statistics
library (gamlss)
library (gamlss.add)
library(gamlss.dist)
library (mgcv)
library (gratia)
library (viridis)
library (refund)
library (cowplot)
library (FoSIntro)
```

# Import

```{r}
pffr_res <- readRDS ("output/pffr_res.RDS")
list2env(pffr_res,globalenv())

df <- readRDS ("output/combined_data.RDS") 
  #filter (study != "lencioni") 

df <- df %>%
  mutate (age2 = case_when(
    age %in% c(3:5) ~ "<6yo",
    age %in% c(6:9) ~ "6-9yo",
    age %in% c(10:19) ~ "10+yo",
    age %in% c(20:29) ~ "20+yo",
    age %in% c(30:39) ~ "30+yo",
    age %in% c(40:49) ~ "40+yo",
    age %in% c(50:59) ~ "50+yo",
    age %in% c(60:69) ~ "60+yo",
    age %in% c(70:79) ~ "70+yo",
    TRUE ~ "80+yo"
  ))
```

# Create new data 

```{r}
new_data <- df %>%
  dplyr::select (age2, ht, speed, strlen) %>%
  group_by(age2) %>%
  summarise(ht = mean (ht, na.rm = TRUE),
            speed = mean (speed, na.rm = TRUE),
            strlen = mean (strlen, na.rm = TRUE)) %>%
  mutate_if (is.numeric, round, 2) %>%
  slice(-9) %>%
  mutate(sex = factor ("m"),
         study = factor("senden"),
         age = c(10, 20, 30, 40, 50, 60, 70, 80, 5))

new_data <- bind_rows(new_data, new_data, new_data)
new_data$speed <- rep (c(0.8, 1.2, 1.4), each = 9)
  
```


# Analysis of ankle power

```{r}

newdata <- new_data %>%
  as.list()
  

pred_ank <- predict (mod_hip, newdata = newdata,
                     type = "response", se.fit = TRUE, exclude = c("s(study)", "sex"))

pred_ank2 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_ank$fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "fit") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric)

pred_ank3 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_ank$se.fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "se") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric,
          fit = pred_ank2$fit,
          lb = fit - 1.96 *se,
          ub = fit + 1.96*se,
          cycle = rep(25:75, nrow(.)/length (25:75)))


#A1: a region of negative power, corresponding to eccentric plantar flexor activity at the ankle during midstance and terminal stance.

a1 <-pred_ank3 %>%
  group_by(age, speed) %>%
  slice_min (fit)  %>%
  mutate (speed = factor (speed))

# A2: a region of positive power, corresponding to the concentric burst of propulsive plantar flexor activity during preswing.
a2 <-pred_ank3 %>%
  group_by(age, speed) %>%
  slice_max (fit)  %>%
  mutate (speed = factor (speed))

```


# Analysis of knee power

```{r}

pred_kne <- predict (mod_kne, newdata = newdata,
                     type = "response", se.fit = TRUE, exclude = c("s(study)", "sex"))

pred_kne2 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_kne$fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "fit") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric)

pred_kne3 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_kne$se.fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "se") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric,
          fit = pred_kne2$fit,
          lb = fit - 1.96 *se,
          ub = fit + 1.96*se)


# K1:  a region of negative power, corresponding to eccentric knee extensor activity at during loading response.
k1 <- pred_kne3 %>%
  filter (cycle <= 20) %>%
  group_by(age, speed) %>%
  slice_min (fit)  %>%
  mutate (speed = factor (speed))

# K2:  a region of positive power, corresponding to concentric knee extensor activity during midstance
k2 <- pred_kne3 %>%
  filter (cycle > 10 & cycle < 40) %>%
  group_by(age, speed) %>%
  slice_max (fit)  %>%
  mutate (speed = factor (speed))


# K3:  a region of negative power, corresponding to eccentric activity in the rectus femoris during preswing
k3 <- pred_kne3 %>%
  filter (cycle > 50 & cycle < 70) %>%
  group_by(age, speed) %>%
  slice_min (fit)  %>%
  mutate (speed = factor (speed))

# K4:   a region of negative power, corresponding to eccentric activity in the hamstrings during terminal swing.
k4 <- pred_kne3 %>%
  filter (cycle > 70) %>%
  group_by(age, speed) %>%
  slice_min (fit)  %>%
  mutate (speed = factor (speed))



```


# Analysis of hip power

```{r}

pred_hip <- predict (mod_hip, newdata = newdata,
                     type = "response", se.fit = TRUE, exclude = c("s(study)", "sex"))

pred_hip2 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_hip$fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "fit") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric)

pred_hip3 <- bind_cols(newdata) %>%
  bind_cols(data.frame (pred_hip$se.fit)) %>%
  pivot_longer(cols = starts_with("X"),
               names_to = "cycle",
               values_to = "se") %>%
  mutate (cycle = str_remove(cycle, "X") %>% as.numeric,
          fit = pred_hip2$fit,
          lb = fit - 1.96 *se,
          ub = fit + 1.96*se)

# H1:  a small region of positive power, not always present, which corresponds to concentric hip extensor activity during loading response.

h1 <- pred_hip3 %>%
  filter (cycle <= 40) %>%
  group_by(age, speed) %>%
  slice_max (fit)  %>%
  mutate (speed = factor (speed))

# H2:  a region of negative power, corresponding to eccentric hip flexor activity during midstance.

h2 <-  pred_hip3 %>%
  filter (cycle > 40 & cycle <60) %>%
  group_by(age, speed) %>%
  slice_min (fit)  %>%
  mutate (speed = factor (speed))

# H3:  a region of positive power, corresponding to concentric activity in the hip flexors during preswing and intial swing.
h3 <- pred_hip3 %>%
  filter (cycle > 50 & cycle < 80) %>%
  group_by(age, speed) %>%
  slice_max (fit)  %>%
  mutate (speed = factor (speed))

```

# Report

## Primary outcomes

```{r}
a2 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

a2 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_max (fit)

a2 %>%
  mutate (age = as.numeric (age)) %>%
  filter (age >3) %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

h1 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)


h3 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

h3 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_max (fit)

```


## Secondary outcomes

```{r}
a1 %>%
  group_by(speed) %>%
  mutate (age = as.numeric (age)) %>%
  filter (age == 2)

a1 %>%
  filter (speed == "0.8") %>%
  mutate (age = as.numeric (age)) %>%
  
  ungroup () %>%
  slice_min (fit)

k2 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

k2 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_max (fit)

k2 %>%
  filter (speed == 0.8) %>%
  ungroup () %>%
  slice_max (fit)

h1 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

h1 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  filter (age == 50)

h1 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  filter (age == 80)

h3 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_min (fit)

h3 %>%
  filter (speed == 1.2) %>%
  ungroup () %>%
  slice_max (fit)

h3 %>%
  filter (speed == 1.4) %>%
  ungroup () %>%
  slice_max (fit)

```

# Report coefficients

```{r}
plot_2Dheatmap_custom <- 
  function (model, 
            plot_ci = TRUE, 
            alpha = 0.05, 
            select, 
            lowCol = "red", 
            highCol = "blue", 
            legend_position = "right", 
            xlab, 
            ylab, 
            base_size = 11, 
            outer_fontsize = 15) 
{

  plotObject <- FoSIntro:::no.plot(model)
  plotObject <- plotObject[[select]]
  plot_data <- expand.grid(plotObject$x, plotObject$y)
  plot_data$fit <- as.vector(plotObject$fit)

  plot_data$ci_lower <- plot_data$fit - qnorm(1 - alpha/2) * 
    as.vector(plotObject$se)
  plot_data$ci_upper <- plot_data$fit + qnorm(1 - alpha/2) * 
    as.vector(plotObject$se)

  xlab <- ifelse(missing(xlab), plotObject$xlab, xlab)
  ylab <- ifelse(missing(ylab), plotObject$ylab, ylab)

  legend_limits <- c(min(plot_data$ci_lower, na.rm = T), 
                            max(plot_data$ci_upper, na.rm = T))

  gg_fit <- ggplot(plot_data) + 
    geom_tile(aes(x = Var1, y = Var2, 
                  fill = fit)) + 
    xlab(xlab) + 
    ylab("") + 
    scale_x_continuous(expand = c(0, 0)) + 
    scale_y_continuous(expand = c(0, 0)) + 
    scale_fill_gradient2(limits = legend_limits,low = lowCol, mid = "white", high = highCol, name = "Joint power (W/kg)") + 
    geom_contour(aes(x = Var1, y = Var2, z = fit), 
                 colour = "darkgrey") + 
    theme_bw(base_size) + 
    theme(plot.title = element_text(hjust = 0.5),legend.position = legend_position)+
    guides (z = "none")
  

    gg_lower <- ggplot(plot_data) + 
      geom_tile(aes(x = Var1, y = Var2, 
                    fill = ci_lower)) + 
      xlab(xlab) + 
      ylab(ylab) + 
      scale_x_continuous(expand = c(0, 0)) + 
      scale_y_continuous(expand = c(0, 0)) + 
      scale_fill_gradient2(limits = legend_limits,low = lowCol, mid = "white", high = highCol, name = "Joint power (W/kg)") + 
      geom_contour(aes(x = Var1, y = Var2, z = ci_lower), 
                   colour = "darkgrey") + 
      theme_bw(base_size) + 
      theme(plot.title = element_text(hjust = 0.5),legend.position = legend_position) +
      guides (z = "none")
    
    gg_upper <- ggplot(plot_data) + 
      geom_tile(aes(x = Var1, y = Var2, 
                    fill = ci_upper)) + 
      xlab(xlab) + 
      ylab("") + 
      scale_x_continuous(expand = c(0, 0)) + 
      scale_y_continuous(expand = c(0, 0)) + 
      scale_fill_gradient2(limits = legend_limits,low = lowCol, mid = "white", high = highCol, name = "Joint power (W/kg)") + 
      geom_contour(aes(x = Var1, y = Var2, z = ci_upper), 
                   colour = "darkgrey") + 
      theme_bw(base_size) + 
      theme(plot.title = element_text(hjust = 0.5),legend.position = legend_position) +
      guides (z = "none")
    
        # arrange the three plots in a single row
        prow <- plot_grid(
          gg_lower + theme(legend.position="none"),
          gg_fit + theme(legend.position="none"),
          gg_upper + theme(legend.position="none"),
          align = 'vh',
          hjust = -1,
          nrow = 1
        )
        
        legend = get_legend(
          gg_upper + theme(legend.box.margin = margin(0, 0, 0, 12))
        )
        
        gg <- plot_grid(prow, legend, rel_widths = c(3, .4))
        
        return(gg)
}

plot_1D_custom <- 
  function (model, 
            plot_ci = TRUE, 
            alpha = 0.05, 
            select, 
            xlab, 
            ylab, 
            base_size = 11, 
            outer_fontsize = 15) 
{

  plotObject <- FoSIntro:::no.plot(model)
  plotObject <- plotObject[[select]]
  plot_data <- expand.grid(plotObject$x, plotObject$y)
  plot_data$fit <- as.vector(plotObject$fit)

  plot_data$ci_lower <- plot_data$fit - qnorm(1 - alpha/2) * 
    as.vector(plotObject$se)
  plot_data$ci_upper <- plot_data$fit + qnorm(1 - alpha/2) * 
    as.vector(plotObject$se)

  xlab <- ifelse(missing(xlab), plotObject$xlab, xlab)
  ylab <- ifelse(missing(ylab), plotObject$ylab, ylab)

  gg <- ggplot(plot_data) + 
    geom_line (aes(x = Var1, y = fit)) + 
    geom_ribbon(aes(x = Var1, ymin = ci_lower, ymax = ci_upper), alpha = 0.4) + 
    xlab(xlab) + 
    ylab(ylab) + 
    theme_bw(base_size) 

  return(gg)
}

```

## Ankle

```{r}
# Ankle
formula (mod_ank)

f3 <- plot_2Dheatmap_custom(mod_ank, 
               select = 3,
               xlab = "Age (years)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f4 <- plot_2Dheatmap_custom(mod_ank, 
               select = 4,
               xlab = "Speed (m/s)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f5 <- plot_2Dheatmap_custom(mod_ank, 
               select = 5,
               xlab = "Age (years)",
               ylab = "Speed (m/s)") +
  theme_cowplot()

# Issues with model terms 6 and 7
# Speed
f6 <- plot_1D_custom(mod_ank, 
               select = 6,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()

# Stride length
f7 <- plot_1D_custom(mod_ank, 
               select = 7,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()
top_row = plot_grid (f3, f4, f5, labels = c("a", "b", "c"), ncol =1)
bottom_row <- plot_grid (f6, f7, labels = c("d", "e"), nrow = 1)

p <- plot_grid(top_row, bottom_row, label_size = 12, 
          rel_heights = c(3, 1), 
          ncol = 1)

tiff ("../manuscript/figsm3.tiff", height = 15, width = 10, res = 200, unit = "in")
  p
dev.off()
```


## Knee

```{r}
formula (mod_kne)

f3 <- plot_2Dheatmap_custom(mod_kne, 
               select = 3,
               xlab = "Age (years)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f4 <- plot_2Dheatmap_custom(mod_kne, 
               select = 4,
               xlab = "Speed (m/s)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f5 <- plot_2Dheatmap_custom(mod_kne, 
               select = 5,
               xlab = "Age (years)",
               ylab = "Speed (m/s)") +
  theme_cowplot()

# Issues with model terms 6 and 7
# Speed
f6 <- plot_1D_custom(mod_kne, 
               select = 6,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()

# Stride length
f7 <- plot_1D_custom(mod_kne, 
               select = 7,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()
top_row = plot_grid (f3, f4, f5, labels = c("a", "b", "c"), ncol =1)
bottom_row <- plot_grid (f6, f7, labels = c("d", "e"), nrow = 1)

p <- plot_grid(top_row, bottom_row, label_size = 12, 
          rel_heights = c(3, 1), 
          ncol = 1)

tiff ("../manuscript/figsm4.tiff", height = 15, width = 10, res = 200, unit = "in")
  p
dev.off()
```


## Hip

```{r}
# Ankle
formula (mod_hip)

f3 <- plot_2Dheatmap_custom(mod_hip, 
               select = 3,
               xlab = "Age (years)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f4 <- plot_2Dheatmap_custom(mod_kne, 
               select = 4,
               xlab = "Speed (m/s)",
               ylab = "Gait cycle (%)") +
  theme_cowplot()

f5 <- plot_2Dheatmap_custom(mod_hip, 
               select = 5,
               xlab = "Age (years)",
               ylab = "Speed (m/s)") +
  theme_cowplot()

# Issues with model terms 6 and 7
# Speed
f6 <- plot_1D_custom(mod_hip, 
               select = 6,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()

# Stride length
f7 <- plot_1D_custom(mod_hip, 
               select = 7,
               xlab = "Gait cycle (%)",
               ylab = "Beta joint power (W/kg)") +
  theme_cowplot()
top_row = plot_grid (f3, f4, f5, labels = c("a", "b", "c"), ncol =1)
bottom_row <- plot_grid (f6, f7, labels = c("d", "e"), nrow = 1)

p <- plot_grid(top_row, bottom_row, label_size = 12, 
          rel_heights = c(3, 1), 
          ncol = 1)

tiff ("../manuscript/figsm5.tiff", height = 15, width = 10, res = 200, unit = "in")
  p
dev.off()
```
